a couple of fixes and utils rename
