+++ lib/twitter.js
@@ -4,8 +4,9 @@ var VERSION = '0.2.8',
   oauth = require('oauth'),
   Cookies = require('cookies'),
   Keygrip = require('keygrip'),
-  streamparser = require('./parser')
-	util = require('./util');
+  streamparser = require('./parser'),
+	util = require('util'),
+	utils = require('./utils');
 
 function Twitter(options) {
   if (!(this instanceof Twitter)) return new Twitter(options);
@@ -38,7 +39,7 @@ function Twitter(options) {
     cookie_options: {},
     cookie_secret: null
   };
-  this.options = util.merge(defaults, options);
+  this.options = utils.merge(defaults, options);
 
   this.oauth = new oauth.OAuth(
     this.options.request_token_url,
@@ -121,9 +122,9 @@ Twitter.prototype.post = function(url, content, content_type, callback) {
   // Workaround: oauth + booleans == broken signatures
   if (content && typeof content === 'object') {
     Object.keys(content).forEach(function(e) {
-      if ( typeof content[e] === 'boolean' )
-        content[e] = content[e].toString();
-    });
+			if ( typeof content[e] === 'boolean' )
+				content[e] = content[e].toString();
+		});
   }
   
   this.oauth.post(url,
@@ -172,7 +173,7 @@ Twitter.prototype.search = function(q, params, callback) {
   }
 
   var url = this.options.search_base + '/search.json';
-  params = util.merge(params, {q:q});
+  params = utils.merge(params, {q:q});
   this.get(url, params, callback);
   return this;
 }
@@ -189,11 +190,11 @@ Twitter.prototype.stream = function(method, params, callback) {
 
   // Iterate on params properties, if any property is an array, convert it to comma-delimited string
   if (params) {
-    Object.keys(params).forEach(function(item) {
-      if (Array.isArray(params[item])) {
-        params[item] = params[item].join(',');
-      }
-    });
+		Object.keys(params).forEach(function(item) {
+			if (util.isArray(params[item])) {
+				params[item] = params[item].join(',');
+			}
+		});
   }
 
   var stream_base = this.options.stream_base,
@@ -286,7 +287,10 @@ Twitter.prototype.stream = function(method, params, callback) {
  */
 Twitter.prototype.cookie = function(req) {
   var keys = null;
-  if ( !this.options.cookie_secret !== null )
+
+	//this make no sense !this.options.cookie_secret return always true or false
+  //if ( !this.options.cookie_secret !== null )
+	if(this.options.cookie_secret)
     keys = new Keygrip(this.options.cookie_secret);
   var cookies = new Cookies( req, null, keys )
   var getState = this.options.getState || function (req, key) {
@@ -314,7 +318,9 @@ Twitter.prototype.login = function(mount, success) {
     this.options.cookie_options.secure = true;
   // Set up the cookie encryption secret if we've been given one
   var keys = null;
-  if ( !this.options.cookie_secret !== null )
+	//the same issue than above
+  //if ( !this.options.cookie_secret !== null )
+	if(this.options.cookie_secret)
     keys = new Keygrip(this.options.cookie_secret);
   // FIXME: ^ so configs that don't use login() won't work?
 
@@ -512,7 +518,7 @@ Twitter.prototype.updateStatus = function(text, params, callback) {
     status: text,
     include_entities: 1
   };
-  params = util.merge(defaults, params);
+  params = utils.merge(defaults, params);
   this.post(url, params, null, callback);
   return this;
 }
@@ -589,7 +595,7 @@ Twitter.prototype.searchUser = function(q, params, callback) {
   }
 
   var url = '/users/search.json';
-  params = util.merge(params, {q:q});
+  params = utils.merge(params, {q:q});
   this.get(url, params, callback);
   return this;
 }
@@ -668,7 +674,7 @@ Twitter.prototype.getLists = function(id, params, callback) {
     defaults.screen_name = id;
   else
     defaults.user_id = id;
-  params = util.merge(defaults, params);
+  params = utils.merge(defaults, params);
 
   var url = '/lists.json';
   this._getUsingCursor(url, params, callback);
@@ -686,7 +692,7 @@ Twitter.prototype.getListMemberships = function(id, params, callback) {
     defaults.screen_name = id;
   else
     defaults.user_id = id;
-  params = util.merge(defaults, params);
+  params = utils.merge(defaults, params);
 
   var url = '/lists/memberships.json';
   this._getUsingCursor(url, params, callback);
@@ -704,7 +710,7 @@ Twitter.prototype.getListSubscriptions = function(id, params, callback) {
     defaults.screen_name = id;
   else
     defaults.user_id = id;
-  params = util.merge(defaults, params);
+  params = utils.merge(defaults, params);
 
   var url = '/lists/subscriptions.json';
   this._getUsingCursor(url, params, callback);
@@ -735,7 +741,7 @@ Twitter.prototype.createList = function(screen_name, list_name, params, callback
   }
 
   var url = '/' + escape(screen_name) + '/lists.json';
-  params = util.merge(params, {name:list_name});
+  params = utils.merge(params, {name:list_name});
   this.post(url, params, null, callback);
   return this;
 }
@@ -766,7 +772,7 @@ Twitter.prototype.getListMembers = function(screen_name, list_id, params, callba
   }
 
   var url = '/' + escape(screen_name) + '/' + escape(list_id) + '/members.json';
-  params = util.merge(params, {key:'users'});
+  params = utils.merge(params, {key:'users'});
   this._getUsingCursor(url, params, callback);
   return this;
 }
@@ -783,7 +789,7 @@ Twitter.prototype.getListSubscribers = function(screen_name, list_id, params, ca
   }
 
   var url = '/' + escape(screen_name) + '/' + escape(list_id) + '/subscribers.json';
-  params = util.merge(params, {key:'users'});
+  params = utils.merge(params, {key:'users'});
   this._getUsingCursor(url, params, callback);
   return this;
 }
@@ -820,7 +826,7 @@ Twitter.prototype.newDirectMessage = function(id, text, params, callback) {
     defaults.screen_name = id;
   else
     defaults.user_id = id;
-  params = util.merge(defaults, params);
+  params = utils.merge(defaults, params);
 
   var url = '/direct_messages/new.json';
   this.post(url, params, null, callback);
@@ -853,7 +859,7 @@ Twitter.prototype.createFriendship = function(id, params, callback) {
     defaults.screen_name = id;
   else
     defaults.user_id = id;
-  params = util.merge(defaults, params);
+  params = utils.merge(defaults, params);
 
   var url = '/friendships/create.json';
   this.post(url, params, null, callback);
@@ -972,7 +978,7 @@ Twitter.prototype.updateProfile = function(params, callback) {
   var defaults = {
     include_entities: 1
   };
-  params = util.merge(defaults, params);
+  params = utils.merge(defaults, params);
 
   var url = '/account/update_profile.json';
   this.post(url, params, null, callback);
@@ -1181,7 +1187,7 @@ Twitter.prototype._getUsingCursor = function(url, params, callback) {
   delete params.key;
 
   // kick off the first request, using cursor -1
-  params = util.merge(params, {cursor:-1});
+  params = utils.merge(params, {cursor:-1});
   this.get(url, params, fetch);
 
   function fetch(err, data) {
